+++ src/presentation/controllers/signup.controller.spec.ts
@@ -1,7 +1,5 @@
 import { SignupController } from './signup.controller';
-import { MissingParamError } from '../errors/missing-param.error';
-import { InvalidParamError } from '../errors/invalid-param.error';
-import { ServerError } from '../errors/server.error';
+import { MissingParamError, InvalidParamError, ServerError } from '../errors';
 import { EmailValidator } from '../protocols/email-validator';
 
 interface SutTypes {
