+++ src/presentation/controllers/signup.controller.ts
@@ -1,9 +1,8 @@
 import { HttpRequest, HttpResponse } from '../protocols/http';
-import { MissingParamError } from '../errors/missing-param.error';
+import { MissingParamError, InvalidParamError } from '../errors';
 import { badRequest, serverError } from '../helpers/http.helper';
 import { Controller } from '../protocols/controller';
 import { EmailValidator } from '../protocols/email-validator';
-import { InvalidParamError } from '../errors/invalid-param.error';
 
 export class SignupController implements Controller {
   private readonly emailValidator: EmailValidator;
