+++ src/presentation/helpers/http.helper.ts
@@ -1,5 +1,5 @@
 import { HttpResponse } from '../protocols/http';
-import { ServerError } from '../errors/server.error';
+import { ServerError } from '../errors';
 
 export const badRequest = (error: Error): HttpResponse => ({
   statusCode: 400,
